WIP refactoring td3pbt
